some fix on tests
