+++ domains/User/index.test.js
@@ -3,7 +3,7 @@ const { hash } = require('bcrypt')
 
 const factory = require('../../utils/helpers/factories')
 const userDomain = require('.')
-const { generatorFakerUser } = require('../../utils/helpers/Faker/user')
+const { generatorFakerUser } = require('../../utils/helpers/fakers')
 const { NotFoundError } = require('../../utils/helpers/errors')
 
 const companyId = 'co_4095e6c0-056d-4b6d-b857-a35584634ad0'
