+++ utils/helpers/factories.js
@@ -1,10 +1,11 @@
 const { factory } = require('factory-girl')
+const { random, date } = require('faker')
 
 const database = require('../../database')
-const { generatorFakerAddress } = require('./Faker/address')
-const { generatorFakerCustomer } = require('./Faker/customer')
-const { generatorFakerUser } = require('./Faker/user')
 const {
+  generatorFakerAddress,
+  generatorFakerCustomer,
+  generatorFakerUser,
   fakerCompany,
   fakerStatus,
   fakerProduct,
@@ -60,6 +61,8 @@ factory.define('order', OrderModel, () => ({
   statusId: factory.assoc('status', 'id', { fakeTransaction: false }),
   customerId: factory.assoc('customer', 'id'),
   userId: factory.assoc('user', 'id'),
+  protocol: random.number(),
+  orderDate: date.future(),
   companyId: 'co_4095e6c0-056d-4b6d-b857-a35584634ad0'
 }))
